+++ ffupdater/src/main/java/de/marmaro/krt/ffupdater/app/impl/Vivaldi.kt
@@ -6,7 +6,6 @@ import de.marmaro.krt.ffupdater.app.AvailableVersionResult
 import de.marmaro.krt.ffupdater.app.BaseAppWithCachedUpdateCheck
 import de.marmaro.krt.ffupdater.app.impl.fetch.ApiConsumer
 import de.marmaro.krt.ffupdater.device.ABI
-import de.marmaro.krt.ffupdater.device.DeviceEnvironment
 
 /**
  * https://vivaldi.com/de/download/
@@ -26,7 +25,12 @@ class Vivaldi : BaseAppWithCachedUpdateCheck() {
     override val signatureHash = "e8a78544655ba8c09817f732768f5689b1662ec4b2bc5a0bc0ec138d33ca3d1e"
 
     override suspend fun updateCheckWithoutCaching(): AvailableVersionResult {
-        val regexPattern = getRegexPattern()
+        val regexPattern = getStringForCurrentAbi(
+            """<a href="(https://downloads.vivaldi.com/stable/Vivaldi.([.0-9]{1,24})_armeabi-v7a.apk)"""",
+            """<a href="(https://downloads.vivaldi.com/stable/Vivaldi.([.0-9]{1,24})_arm64-v8a.apk)"""",
+            null,
+            """<a href="(https://downloads.vivaldi.com/stable/Vivaldi.([.0-9]{1,24})_x86-64.apk)"""",
+        )
         val content = ApiConsumer.consumeNetworkResource(DOWNLOAD_WEBSITE_URL, String::class)
         val regexMatch = Regex(regexPattern).find(content)
         checkNotNull(regexMatch) { "Can't find download link with regex pattern '$regexPattern'." }
@@ -44,24 +48,7 @@ class Vivaldi : BaseAppWithCachedUpdateCheck() {
         )
     }
 
-    private fun getRegexPattern(): String {
-        return DeviceEnvironment.abis.mapNotNull {
-            when (it) {
-                ABI.ARM64_V8A -> REGEX_PATTERN_ARM64_V8A
-                ABI.ARMEABI_V7A -> REGEX_PATTERN_ARMEABI_V7A
-                ABI.X86_64 -> REGEX_PATTERN_X86_64
-                ABI.X86, ABI.ARMEABI, ABI.MIPS, ABI.MIPS64 -> null
-            }
-        }.first()
-    }
-
     companion object {
         const val DOWNLOAD_WEBSITE_URL = "https://vivaldi.com/download/"
-        const val REGEX_PATTERN_ARMEABI_V7A =
-            """<a href="(https://downloads.vivaldi.com/stable/Vivaldi.([.0-9]{1,24})_armeabi-v7a.apk)""""
-        const val REGEX_PATTERN_ARM64_V8A =
-            """<a href="(https://downloads.vivaldi.com/stable/Vivaldi.([.0-9]{1,24})_arm64-v8a.apk)""""
-        const val REGEX_PATTERN_X86_64 =
-            """<a href="(https://downloads.vivaldi.com/stable/Vivaldi.([.0-9]{1,24})_x86-64.apk)""""
     }
 }